Update draft_submit option to be present in mod_assign form
This option was only being inserted into the mod_assignment form, as we
were checking for the presence of a field called "var4" which does not
exist in mod_assign.  This has been switched to submissiondrafts, which is
also now used to toggle the disabled state of the draft_submit field.