Submit files for assessable_submitted events too
The mod_assign module triggers an assessable_submitted event, which we
were listening for but not processing correctly.  It should be handled the
same way as file_done, as both result in no pathnamehashes being provided,
but there are some differences between how mod_assign and mod_assignment
handle these events.